requirements.txt
asgiref==3.8.1
black==24.3.0
click==8.1.7
Django==5.0.4
djangorestframework
djangorestframework-jwt-5
jedi==0.19.1
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pudb==2024.1
Pygments==2.17.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.11.0
urwid==2.6.10
urwid_readline==0.14
wcwidth==0.2.13